refator : 인증관련 기능 리팩터링
