+++ src/main/java/nextstep/subway/auth/ui/authentication/SessionAuthenticationInterceptor.java
@@ -1,14 +1,28 @@
 package nextstep.subway.auth.ui.authentication;
 
+import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.infrastructure.SecurityContext;
+import nextstep.subway.auth.ui.converter.AuthenticationConverter;
+import nextstep.subway.member.application.CustomUserDetailsService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
+import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
+
 public class SessionAuthenticationInterceptor extends AuthenticationInterceptor{
+
+    public SessionAuthenticationInterceptor(AuthenticationConverter authenticationConverter, UserDetailService userDetailService) {
+        super(authenticationConverter, userDetailService);
+    }
+
     @Override
     public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-
+        HttpSession httpSession = request.getSession();
+        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
+        response.setStatus(HttpServletResponse.SC_OK);
     }
 }
