+++ src/main/java/nextstep/subway/auth/ui/authentication/TokenAuthenticationInterceptor.java
@@ -1,6 +1,14 @@
 package nextstep.subway.auth.ui.authentication;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.dto.TokenResponse;
+import nextstep.subway.auth.infrastructure.JwtTokenProvider;
+import nextstep.subway.auth.ui.converter.AuthenticationConverter;
+import nextstep.subway.member.application.CustomUserDetailsService;
+import org.springframework.http.MediaType;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -8,8 +16,34 @@ import java.io.IOException;
 
 public class TokenAuthenticationInterceptor extends AuthenticationInterceptor{
 
+    AuthenticationConverter authenticationConverter;
+    CustomUserDetailsService userDetailsService;
+    JwtTokenProvider jwtTokenProvider;
+
+    public TokenAuthenticationInterceptor(AuthenticationConverter authenticationConverter, UserDetailService userDetailService, JwtTokenProvider jwtTokenProvider) {
+        super(authenticationConverter, userDetailService);
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
     @Override
     public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        // 토큰 생성및 응답객체에 세팅
+        TokenResponse tokenResponse = getTokenResponse(authentication);
+        setTokenResponse(response, tokenResponse);
+    }
 
+    private void setTokenResponse(HttpServletResponse response, TokenResponse tokenResponse) throws IOException {
+        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.getOutputStream().print(responseToClient);
+    }
+
+    private TokenResponse getTokenResponse(Authentication authentication) throws JsonProcessingException {
+        // Authentication -> TokenResponse
+        String payload = new ObjectMapper().writeValueAsString(authentication.getPrincipal());
+        String token = jwtTokenProvider.createToken(payload);
+        TokenResponse tokenResponse = new TokenResponse(token);
+        return tokenResponse;
     }
 }
