+++ src/main/java/nextstep/subway/auth/ui/converter/SessionAuthenticationConverter.java
@@ -1,12 +1,23 @@
 package nextstep.subway.auth.ui.converter;
 
 import nextstep.subway.auth.domain.AuthenticationToken;
+import org.springframework.stereotype.Component;
 
 import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+import java.util.Map;
 
+@Component
 public class SessionAuthenticationConverter implements AuthenticationConverter {
+    public static final String USERNAME_FIELD = "username";
+    public static final String PASSWORD_FIELD = "password";
+
     @Override
-    public AuthenticationToken convert(HttpServletRequest request) {
-        return null;
+    public AuthenticationToken convert(HttpServletRequest request) throws IOException {
+        Map<String, String[]> paramMap = request.getParameterMap();
+        String principal = paramMap.get(USERNAME_FIELD)[0];
+        String credentials = paramMap.get(PASSWORD_FIELD)[0];
+
+        return new AuthenticationToken(principal, credentials);
     }
 }
