+++ src/main/java/nextstep/subway/auth/ui/persistence/SessionSecurityContextInterceptor.java
@@ -1,4 +1,27 @@
 package nextstep.subway.auth.ui.persistence;
 
+import nextstep.subway.auth.infrastructure.SecurityContext;
+import nextstep.subway.auth.infrastructure.SecurityContextHolder;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
+
 public class SessionSecurityContextInterceptor extends SecurityContextInterceptor {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        if (SecurityContextHolder.getContext().getAuthentication() != null) {
+            return true;
+        }
+
+        SecurityContext securityContext = (SecurityContext) request.getSession().getAttribute(SPRING_SECURITY_CONTEXT_KEY);
+        if (securityContext != null) {
+            SecurityContextHolder.setContext(securityContext);
+        }
+        return true;
+    }
+
+
 }
