--- src/main/java/nextstep/subway/auth/ui/session/SessionSecurityContextPersistenceInterceptor.java
@@ -1,30 +1,0 @@
-package nextstep.subway.auth.ui.session;
-
-import nextstep.subway.auth.infrastructure.SecurityContext;
-import nextstep.subway.auth.infrastructure.SecurityContextHolder;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
-
-public class SessionSecurityContextPersistenceInterceptor implements HandlerInterceptor {
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        if (SecurityContextHolder.getContext().getAuthentication() != null) {
-            return true;
-        }
-
-        SecurityContext securityContext = (SecurityContext) request.getSession().getAttribute(SPRING_SECURITY_CONTEXT_KEY);
-        if (securityContext != null) {
-            SecurityContextHolder.setContext(securityContext);
-        }
-        return true;
-    }
-
-    @Override // view 랜더링 이후
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
-        SecurityContextHolder.clearContext();
-    }
-}
