+++ src/main/java/nextstep/subway/member/application/CustomUserDetailsService.java
@@ -1,12 +1,13 @@
 package nextstep.subway.member.application;
 
+import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomUserDetailsService {
+public class CustomUserDetailsService implements UserDetailService {
     private MemberRepository memberRepository;
 
     public CustomUserDetailsService(MemberRepository memberRepository) {
