+++ src/test/java/nextstep/subway/auth/ui/converter/AuthenticationConverterTest.java
@@ -1,13 +1,10 @@
 package nextstep.subway.auth.ui.converter;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.subway.auth.domain.AuthenticationToken;
 import nextstep.subway.auth.dto.TokenRequest;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpRequest;
 import org.springframework.mock.web.MockHttpServletRequest;
 
 import java.io.IOException;
